fix: apply review findings to distribution code
- install.sh: validate MATRIX_REPO/MATRIX_VERSION env vars, fix
  tag/asset name mismatch (v-prefixed tag for URL, stripped for .pkg)
- release.yml: add typecheck gate before publish
- cli/index.ts: read version from package.json instead of hardcoded 0.0.1
- bin/matrix.mjs: use shared find-tsx-loader.mjs
